re #BITBUCKET-180 moving raster symbology to a separate package and adding a test Branch: feature/GEOPROCESSING-180-raster-calculator-expressions
